Changes and comments fro actual changes
* SliceDisptacher is actually a coherence concept not restricted to MSI, right?
* The number of LLC is needed and is therrefore provide by template.
* Some comment for further modification suggestions.
